fix(tabs overlap): fix android version and messages
