+++ app/[id]/(tabs)/_layout.tsx
@@ -1,5 +1,4 @@
-
-import { NativeTabs } from "expo-router/build/native-tabs";
+import { NativeTabs } from "expo-router/unstable-native-tabs";
 
 export default function ItemDetailsLayout() {
